+++ libs/base_macros/src/bodies/mod.rs
@@ -9,7 +9,9 @@ crate::items! {
 
     mod shared;
 
-    mod compile; // cif, compile[_attr|_doc]
-    mod ident; // coalesce, field_of, ident_total, ident_total_unique, ident_unique
-    pub(crate) use {compile::*, ident::*};
+    mod compile; // cif!, compile[_attr|_doc]!
+    mod ident; // coalesce!, field_of!, ident_total!, ident_total_unique!, ident_unique!
+    mod repeat; // repeat!
+
+    pub(crate) use {compile::*, ident::*, repeat::*};
 }